server/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
django-cors-headers>=4.0
django-filter>=23.1
django-anymail>=9.0
python-dotenv>=1.0
Pillow>=10
gunicorn>=20
requests>=2.31.0

# MailerSend SDK (used by accounts password reset)
mailersend>=2.0.0

# Optional / dev packages (install locally if required)
# uvicorn>=0.27.1
# pipenv>=2023.12.1
